feat: redirect when already logged in
